LayoutTests/crypto/webkitSubtle/hmac-postMessage-expected.txt

CONSOLE MESSAGE: line 26: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test sending crypto keys via postMessage.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'HMAC'
PASS key.algorithm.length is 128
PASS key.usages is ["sign", "verify"]
PASS successfullyParsed is true

TEST COMPLETE
